Add renderer API

+++ Engine/src/Engine/Renderer/Buffer.h
@@ -132,6 +132,6 @@ namespace Engine
 		virtual void Unbind() const = 0;
 		virtual uint32_t GetCount() const = 0;
 
-		static IndexBuffer* Create(uint32_t* indices, size_t count);
+		static IndexBuffer* Create(uint32_t* indices, uint32_t count);
 	};
 }
